implementing code review
