+++ src/main/java/org/influxdb/BatchOptions.java
@@ -1,8 +1,8 @@
 package org.influxdb;
 
 import org.influxdb.dto.Point;
-import org.influxdb.impl.BatchOptionsImpl;
 
+import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadFactory;
 import java.util.function.BiConsumer;
 
@@ -10,21 +10,51 @@ import java.util.function.BiConsumer;
  * BatchOptions are used to configure batching of individual data point writes
  * into InfluxDB. See {@link InfluxDB#enableBatch(BatchOptions)}
  */
-public interface BatchOptions {
+public final class BatchOptions implements Cloneable {
 
-  BatchOptions DEFAULTS = BatchOptionsImpl.DEFAULTS;
+  /**
+   * Default batch options. This class is immutable, each configuration
+   * is built by taking the DEFAULTS and setting specific configuration
+   * properties.
+   */
+  public static final BatchOptions DEFAULTS = new BatchOptions();
+
+  // default values here are consistent with Telegraf
+  public static final int DEFAULT_BATCH_ACTIONS_LIMIT = 1000;
+  public static final int DEFAULT_BATCH_INTERVAL_DURATION = 1000;
+  public static final int DEFAULT_JITTER_INTERVAL_DURATION = 0;
+
+  private int actions = DEFAULT_BATCH_ACTIONS_LIMIT;
+  private int flushDuration = DEFAULT_BATCH_INTERVAL_DURATION;
+  private int jitterDuration = DEFAULT_JITTER_INTERVAL_DURATION;
+
+  private ThreadFactory threadFactory = Executors.defaultThreadFactory();
+  BiConsumer<Iterable<Point>, Throwable> exceptionHandler = (points, throwable) -> {
+  };
+  private InfluxDB.ConsistencyLevel consistency = InfluxDB.ConsistencyLevel.ONE;
+
+  private BatchOptions() {
+  }
 
   /**
    * @param actions the number of actions to collect
    * @return the BatchOptions instance to be able to use it in a fluent manner.
    */
-  BatchOptions actions(final int actions);
+  public BatchOptions actions(final int actions) {
+    BatchOptions clone = getClone();
+    clone.actions = actions;
+    return clone;
+  }
 
   /**
    * @param flushDuration the time to wait at most (milliseconds).
    * @return the BatchOptions instance to be able to use it in a fluent manner.
    */
-  BatchOptions flushDuration(final int flushDuration);
+  public BatchOptions flushDuration(final int flushDuration) {
+    BatchOptions clone = getClone();
+    clone.flushDuration = flushDuration;
+    return clone;
+  }
 
   /**
    * Jitters the batch flush interval by a random amount. This is primarily to avoid
@@ -34,57 +64,92 @@ public interface BatchOptions {
    * @param jitterDuration (milliseconds)
    * @return the BatchOptions instance to be able to use it in a fluent manner.
    */
-  BatchOptions jitterDuration(final int jitterDuration);
+  public BatchOptions jitterDuration(final int jitterDuration) {
+    BatchOptions clone = getClone();
+    clone.jitterDuration = jitterDuration;
+    return clone;
+  }
 
   /**
    * @param threadFactory a ThreadFactory instance to be used
    * @return the BatchOptions instance to be able to use it in a fluent manner.
    */
-  BatchOptions threadFactory(final ThreadFactory threadFactory);
+  public BatchOptions threadFactory(final ThreadFactory threadFactory) {
+    BatchOptions clone = getClone();
+    clone.threadFactory = threadFactory;
+    return clone;
+  }
 
   /**
    * @param exceptionHandler a consumer function to handle asynchronous errors
    * @return the BatchOptions instance to be able to use it in a fluent manner.
    */
-  BatchOptions exceptionHandler(final BiConsumer<Iterable<Point>, Throwable> exceptionHandler);
+  public BatchOptions exceptionHandler(final BiConsumer<Iterable<Point>, Throwable> exceptionHandler) {
+    BatchOptions clone = getClone();
+    clone.exceptionHandler = exceptionHandler;
+    return clone;
+  }
 
   /**
    * @param consistency cluster consistency setting (how many nodes have to store data points
-   * to treat a write as a success)
+   *                    to treat a write as a success)
    * @return the BatchOptions instance to be able to use it in a fluent manner.
    */
-  BatchOptions setConsistency(final InfluxDB.ConsistencyLevel consistency);
+  public BatchOptions setConsistency(final InfluxDB.ConsistencyLevel consistency) {
+    BatchOptions clone = getClone();
+    clone.consistency = consistency;
+    return clone;
+  }
 
-
-    /**
-     * @return actions the number of actions to collect
-     */
-  int getActions();
+  /**
+   * @return actions the number of actions to collect
+   */
+  public int getActions() {
+    return actions;
+  }
 
   /**
    * @return flushDuration the time to wait at most (milliseconds).
    */
-  int getFlushDuration();
+  public int getFlushDuration() {
+    return flushDuration;
+  }
 
   /**
    * @return batch flush interval jitter value (milliseconds)
    */
-  int getJitterDuration();
+  public int getJitterDuration() {
+    return jitterDuration;
+  }
 
   /**
    * @return a ThreadFactory instance to be used
    */
-  ThreadFactory getThreadFactory();
+  public ThreadFactory getThreadFactory() {
+    return threadFactory;
+  }
 
   /**
    * @return a consumer function to handle asynchronous errors
    */
-  BiConsumer<Iterable<Point>, Throwable> getExceptionHandler();
+  public BiConsumer<Iterable<Point>, Throwable> getExceptionHandler() {
+    return exceptionHandler;
+  }
 
   /**
    * @return cluster consistency setting (how many nodes have to store data points
    * to treat a write as a success)
    */
-  InfluxDB.ConsistencyLevel getConsistency();
+  public InfluxDB.ConsistencyLevel getConsistency() {
+    return consistency;
+  }
+
+  private BatchOptions getClone() {
+    try {
+      return (BatchOptions) this.clone();
+    } catch (CloneNotSupportedException e) {
+      throw new RuntimeException(e);
+    }
+  }
 
 }
